+++ plugins/Profile/src/Controller/ProfilesController.php
@@ -7,7 +7,8 @@ use Cake\Event\Event;
 use Cake\ORM\Table;
 use Cake\ORM\Query;
 use Cake\ORM\TableRegistry;
-
+use Cake\Utility\Inflector;
+use Cake\Routing\Router;
 use App\Controller\AppController;
 
 class ProfilesController extends AppController
@@ -27,6 +28,7 @@ class ProfilesController extends AppController
         'StaffSubjects',
         'StaffBehaviours',
         'Licenses',
+        'StaffAttendances',
     ];
 
     public function initialize()
@@ -58,7 +60,6 @@ class ProfilesController extends AppController
             'StaffSections'         => ['className' => 'Staff.StaffSections', 'actions' => ['index', 'view']],
             'StaffClasses'          => ['className' => 'Staff.StaffClasses', 'actions' => ['index', 'view']],
             'StaffQualifications'   => ['className' => 'Staff.Qualifications'],
-            'StaffAbsences'         => ['className' => 'Staff.Absences', 'actions' => ['index', 'view']],
             'StaffExtracurriculars' => ['className' => 'Staff.Extracurriculars'],
             'TrainingResults'       => ['className' => 'Staff.TrainingResults', 'actions' => ['index', 'view']],
 
@@ -110,7 +111,9 @@ class ProfilesController extends AppController
     public function StaffBehaviours()         { $this->ControllerAction->process(['alias' => __FUNCTION__, 'className' => 'Staff.StaffBehaviours']); }
     public function StudentOutcomes()         { $this->ControllerAction->process(['alias' => __FUNCTION__, 'className' => 'Student.StudentOutcomes']); }
     public function ScholarshipApplications() { $this->ControllerAction->process(['alias' => __FUNCTION__, 'className' => 'Profile.ScholarshipApplications']); }
-    public function Demographic()            { $this->ControllerAction->process(['alias' => __FUNCTION__, 'className' => 'User.Demographic']); }
+    public function Demographic()             { $this->ControllerAction->process(['alias' => __FUNCTION__, 'className' => 'User.Demographic']); }
+    public function ProfileStudents()         { $this->ControllerAction->process(['alias' => __FUNCTION__, 'className' => 'Profile.Students']); }
+    public function ProfileStudentUser()      { $this->ControllerAction->process(['alias' => __FUNCTION__, 'className' => 'Profile.StudentUser']); }
 
     // health
     public function Healths()               { $this->ControllerAction->process(['alias' => __FUNCTION__, 'className' => 'Health.Healths']); }
@@ -122,7 +125,7 @@ class ProfilesController extends AppController
     public function HealthMedications()     { $this->ControllerAction->process(['alias' => __FUNCTION__, 'className' => 'Health.Medications']); }
     public function HealthTests()           { $this->ControllerAction->process(['alias' => __FUNCTION__, 'className' => 'Health.Tests']); }
     // End Health
-    
+
     // Special Needs
     public function SpecialNeedsReferrals()   { $this->ControllerAction->process(['alias' => __FUNCTION__, 'className' => 'SpecialNeeds.SpecialNeedsReferrals']); }
     public function SpecialNeedsAssessments() { $this->ControllerAction->process(['alias' => __FUNCTION__, 'className' => 'SpecialNeeds.SpecialNeedsAssessments']); }
@@ -132,6 +135,26 @@ class ProfilesController extends AppController
     // Special Needs - End
 
     public function Employments()           { $this->ControllerAction->process(['alias' => __FUNCTION__, 'className' => 'User.UserEmployments']); }
+
+    public function HistoricalStaffLeave()
+    {
+        $this->ControllerAction->process(['alias' => __FUNCTION__, 'className' => 'Historical.HistoricalStaffLeave']);
+    }
+    // AngularJS
+    public function StaffAttendances()
+    {
+        $institutionId = null;
+        $staffId = $this->Auth->user('id');
+        $tabElements = $this->getCareerTabElements();
+
+        $crumbTitle = __(Inflector::humanize(Inflector::underscore($this->request->param('action'))));
+        $this->Navigation->addCrumb($crumbTitle);
+        $this->set('institution_id', $institutionId);
+        $this->set('staff_id', $staffId);
+        $this->set('tabElements', $tabElements);
+        $this->set('selectedAction', 'Attendances');
+        $this->set('ngController', 'StaffAttendancesCtrl as $ctrl');
+    }
     // End
 
     public function implementedEvents()
@@ -196,6 +219,12 @@ class ProfilesController extends AppController
                     'alert.svc',
                     'student.examination_results.ctrl',
                     'student.examination_results.svc'
+                ]);
+                break;
+            case 'StaffAttendances':
+                $this->Angular->addModules([
+                    'staff.attendances.ctrl',
+                    'staff.attendances.svc'
                 ]);
                 break;
         }
@@ -321,6 +350,10 @@ class ProfilesController extends AppController
                 $idKey[$model->aliasField('student_id')] = $userId;
                 $exists = $model->exists($idKey);
 
+               if (in_array($model->alias(), ['Students'])) {
+                    $params[$model->aliasField('guardian_id')] = $userId;
+                    $exists = $model->exists($params);
+                }
                 /**
                  * if the sub model's id does not belongs to the main model through relation, redirect to sub model index page
                  */
@@ -480,8 +513,8 @@ class ProfilesController extends AppController
             'Positions' => ['text' => __('Positions')],
             'Classes' => ['text' => __('Classes')],
             'Subjects' => ['text' => __('Subjects')],
-            'Absences' => ['text' => __('Absences')],
             'Leave' => ['text' => __('Leave')],
+            'Attendances' => ['text' => __('Attendances')],
             'Behaviours' => ['text' => __('Behaviours')],
             'Appraisals' => ['text' => __('Appraisals')],
         ];